Change useItemFromHandOnObject parameter's type from int to long (this is the actual type of Gobs' IDs)

+++ src/haven/Game.java
@@ -832,7 +832,7 @@ public class Game extends Widget {
 
     }
 
-    public boolean useItemFromHandOnObject(int id) {
+    public boolean useItemFromHandOnObject(long id) {
 
         GameUI gui = gameui();
